Don't change SAC files with no "b" or reftime
Related to issue 1285, and test file "testxy.sac".  If there are relative time
headers, even if there's a "b", throw an exception. If there's a "b" and no
relative times, use the "b" and write the "e" and nothing else. If there's no
"b" and no relative times, write the SAC file like it's an "ib" type file and
proceed.